Refactored conversion logic structure described in PR
