+++ service/src/main/java/com/epam/rft/atsy/service/configuration/ConverterConfiguration.java
@@ -5,13 +5,13 @@ import com.epam.rft.atsy.persistence.repositories.ChannelRepository;
 import com.epam.rft.atsy.persistence.repositories.PositionRepository;
 import com.epam.rft.atsy.persistence.repositories.StatesHistoryRepository;
 import com.epam.rft.atsy.service.ConverterService;
-import com.epam.rft.atsy.service.converter.ApplicationTwoWayConverter;
+import com.epam.rft.atsy.service.converter.impl.ApplicationTwoWayConverter;
 import com.epam.rft.atsy.service.converter.BaseConverter;
-import com.epam.rft.atsy.service.converter.CandidateApplicationOneWayConverter;
+import com.epam.rft.atsy.service.converter.impl.CandidateApplicationOneWayConverter;
 import com.epam.rft.atsy.service.converter.ConverterAdapter;
-import com.epam.rft.atsy.service.converter.StateFlowTwoWayConverter;
-import com.epam.rft.atsy.service.converter.StateHistoryTwoWayConverter;
-import com.epam.rft.atsy.service.converter.StateHistoryViewTwoWayConverter;
+import com.epam.rft.atsy.service.converter.impl.StateFlowTwoWayConverter;
+import com.epam.rft.atsy.service.converter.impl.StateHistoryTwoWayConverter;
+import com.epam.rft.atsy.service.converter.impl.StateHistoryViewTwoWayConverter;
 import com.epam.rft.atsy.service.impl.ModelMapperConverterServiceImpl;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
