+++ service/src/main/java/com/epam/rft/atsy/service/converter/OneWayConverter.java
@@ -7,15 +7,15 @@ import com.epam.rft.atsy.service.ConverterService;
  * If you need to create a new OneWayConverter implementation, you should extend the
  * {@link AbstractOneWayConverter} class, because {@link BaseConverter#generate()} is already
  * implemented in it.
- * @param <S> the source type
- * @param <T> the target type
+ * @param <F> the first type
+ * @param <S> the second type
  */
-public interface OneWayConverter<S, T> extends BaseConverter {
+public interface OneWayConverter<F, S> extends BaseConverter {
   /**
-   * Converts the {@code source} with {@code S} type to a {@code T} type object.
+   * Converts the {@code source} with {@code F} type to a {@code S} type object.
    * @param source the object to convert
    * @return the converted object
    */
-  T sourceTypeToTargetType(S source);
+  S firstTypeToSecondType(F source);
 
 }
